pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-thumbnail"
version = "0.1.0"
description = "Utilities for extracting headshots and composing podcast thumbnails."
authors = [{ name = "Adi" }]
requires-python = ">=3.9"
dependencies = [
    "pillow>=10.4",
    "google-genai>=1.44.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.7",
]

[project.scripts]
podthumb = "orchestration_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
